rename run.rs to app.rs and move the App struct into it
